extract remotes
